crate for buffer and parser
