+++ Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [workspace]
-members = ["wf_common", "wf_validation", "wf_codec", "wf_field", "wf_crypto"]
+members = ["wf_common", "wf_validation", "wf_codec", "wf_field", "wf_crypto", "wf_buffer"]
 
 
 [dependencies]
@@ -23,3 +23,5 @@ wf_codec = { path = "wf_codec" }
 wf_field = { path = "wf_field" }
 wf_crypto = { path = "wf_crypto" }
 wf_auth = { path = "wf_auth" }
+wf_buffer = { path = "wf_buffer" }
+wf_parser = {path = "wf_parser"}
