+++ src/lib.rs
@@ -2,11 +2,9 @@ use error::WhiteflagError;
 
 mod error;
 mod wf_account;
-mod wf_buffer;
 #[allow(dead_code)]
 mod wf_core;
 mod wf_json;
-mod wf_parser;
 
 pub fn encode_from_json<T: AsRef<str>>(json: T) -> Result<String, WhiteflagError> {
     let message: wf_json::WhiteflagFieldValues =
