+++ wf_buffer/src/encode.rs
