+++ wf_field/src/lib.rs
@@ -23,6 +23,9 @@ pub const FIELD_VERSION: &'static str = "Version";
 pub const FIELD_MESSAGETYPE: &'static str = "MessageCode";
 pub const FIELD_TESTMESSAGETYPE: &'static str = "PseudoMessageCode";
 
+pub trait FieldValue: AsRef<str> + Into<String> + std::fmt::Debug {}
+impl<T> FieldValue for T where T: AsRef<str> + Into<String> + std::fmt::Debug {}
+
 impl From<&Field> for Vec<u8> {
     fn from(field: &Field) -> Self {
         field.encode()
