+++ wf_parser/src/lib.rs
@@ -1,0 +1,53 @@
+#[cfg(test)]
+mod test;
+
+mod message_code_parser;
+mod message_header_parser;
+mod parsed_field_definition;
+mod wf_header;
+
+pub use message_code_parser::MessageCodeParser;
+pub use message_header_parser::MessageHeaderParser;
+pub use parsed_field_definition::ParsedFieldDefinition;
+pub use wf_header::{MessageHeaderFields, MessageHeaderValues};
+
+use wf_field::{definitions::convert_value_to_code, Field};
+
+pub trait MessageHeader {
+    type Target: ?Sized;
+
+    fn prefix(&self) -> &Self::Target;
+    fn version(&self) -> &Self::Target;
+    fn encryption_indicator(&self) -> &Self::Target;
+    fn duress_indicator(&self) -> &Self::Target;
+    fn message_code(&self) -> &Self::Target;
+    fn reference_indicator(&self) -> &Self::Target;
+    fn referenced_message(&self) -> &Self::Target;
+}
+
+#[repr(usize)]
+#[derive(Copy, Clone)]
+pub enum MessageHeaderOrder {
+    Prefix = 0,
+    Version = 1,
+    EncryptionIndicator = 2,
+    DuressIndicator = 3,
+    MessageCode = 4,
+    ReferenceIndicator = 5,
+    ReferencedMessage = 6,
+}
+
+impl<'a> MessageHeaderOrder {
+    pub fn as_usize(&self) -> usize {
+        *self as usize
+    }
+
+    pub fn get<'b>(&'a self, fields: &'b [Field]) -> &'b Field {
+        &fields[self.as_usize()]
+    }
+
+    pub fn get_code(fields: &[Field]) -> (&Field, char) {
+        let field = Self::MessageCode.get(fields);
+        (field, convert_value_to_code(field.get()))
+    }
+}
