+++ wf_parser/src/message_code_parser.rs
@@ -1,9 +1,7 @@
-use crate::wf_core::FieldValue;
-
 use super::{MessageHeader, MessageHeaderOrder, MessageHeaderParser};
 use wf_field::{
     definitions::{convert_value_to_code, get_body_from_code_char},
-    FieldDefinition,
+    FieldDefinition, FieldValue,
 };
 
 #[derive(Debug)]
