+++ wf_parser/src/wf_header.rs
@@ -1,4 +1,4 @@
-use crate::wf_buffer::WhiteflagBuffer;
+use wf_buffer::WhiteflagBuffer;
 use wf_field::{
     definitions,
     definitions::{convert_value_to_code, get_body_from_code},
